+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v3.1.0
 
 * Added `DistributedJob::Job#push_all`
+* Added `DistributedJob::Job#open_part?`
 
 ## v3.0.1